+++ client/src/components/mapsAndCharts/SimpleMap.js
