+++ client/src/components/pages/Home.js
@@ -3,14 +3,14 @@ import React, { useState } from "react";
 // import Form from "../card/AuthPost";
 import Danger from "../Danger";
 import Search from "./Search";
-import Chart from "../Chart";
-import BarChart from "../BarChart";
+import Chart from "../mapsAndCharts/Chart";
+import BarChart from "../mapsAndCharts/BarChart";
 import { Button } from "@material-ui/core";
 import Loading from "../Loading";
 import API from "../../utils/API";
 import { useUserContext } from "../context/UserContext";
 import Weather from "../Weather";
-import MyMap from "../MyMap";
+import MyMap from "../mapsAndCharts/MyMap";
 import Earthquake from "../Earthquake";
 import "./Home.css";
 const maxDays = 60;
